+++ src/plugins/developer.sh
@@ -46,11 +46,11 @@ print_developer_include_sh()
 }
 
 
-print_developer_environment_all_sh()
+print_developer_environment_global_sh()
 {
-   log_entry "print_developer_environment_all_sh" "$@"
+   log_entry "print_developer_environment_global_sh" "$@"
 
-   print_minimal_environment_all_sh "$@"
+   print_minimal_environment_global_sh "$@"
 }